% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociogram.R
\name{actorSociogram}
\alias{actorSociogram}
\title{Sociogram with actor information}
\usage{
actorSociogram(smatrix)
}
\arguments{
\item{smatrix}{sociomatrix}
}
\value{
Circular plot with actor information
}
\description{
Function to obtain the circle diagram with actor relationship from an sociomatrix.
}
\details{
The function actorSociogram is obtained by the sociometric matrix.
Return a circular diagram with actor relationship between the animals, where
the arrow shows the direction of the relationship and the line thickness
indicates the number of encounters (the thicker the line, the greater
the number of encounters).
}
\examples{
"There is no example"

NULL

}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
