% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barDom}
\alias{barDom}
\title{Barplot from the variables obtained in the dvalue}
\usage{
barDom(dvalue, variable)
}
\arguments{
\item{dvalue}{Dominance value}

\item{variable}{the column with social hierarchy or social rank information.}
}
\value{
histogram of social dominance
}
\description{
Generates a barplot from the variables obtained in the dvalue
function (dominance value, social hierarchy and social rank)
}
\details{
It is a function that plots the social hierarchy or social rank information of a group
in a barplot. The function uses the options provided by ggplot2.
}
\examples{
"There is no example"
NULL

}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
