% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bpDom}
\alias{bpDom}
\title{Boxplot}
\usage{
bpDom(y, x)
}
\arguments{
\item{y}{the column with animal information.}

\item{x}{the column with social hierarchy or social rank information.}
}
\value{
boxplot of social hierarchy or social rank
}
\description{
Function to obtain the boxplot of social hierarchy categories
from dvalue function.
}
\details{
It is a simple function that plots the social hierarchy or social rank information of a group
in a boxplot. The function uses the options provided by ggplot2.
}
\examples{
"There is no example"

NULL

}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
