% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{freqActor}
\alias{freqActor}
\title{Frequency of an animal was actor}
\usage{
freqActor(x)
}
\arguments{
\item{x}{dataset with replacements information.}
}
\value{
Frequency of an animal was an actor
}
\description{
Function to identify frequency that one animal was actor regarding the herd at bins.
}
\details{
freqActor is only applied for dataset with columns ordained and named as follows: actor and reactor.
}
\examples{
x <- data.frame(actor = c(6,3,5,4,2,1,3,5,8,6,9,3,2,1,1),
               reactor = c(2,6,3,5,4,2,1,3,5,1,6,4,3,2,5))                
                                   
freqA <- freqActor(x)                                                
                                                           
print(freqA)

}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
