% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{improved_index}
\alias{improved_index}
\title{Improved linearity index}
\usage{
improved_index(dmatrix, smatrix)
}
\arguments{
\item{dmatrix}{dyatic matrix}

\item{smatrix}{sociomatrix}
}
\value{
h' index
}
\description{
Function to obtain the linearity index improved by de Vries (1995).
}
\details{
The function improved_index is obtained by the following expression: h' = h(6/(n^3-n)*u),
where "h'" is the linearity index, "n" is the total of animals, "u" is the unknown or
tied relationships.
}
\references{
de Vries, H. (1995). An improved test of linearity in dominance hierarchies
containing unknown or tied relationships. Animal Behaviour, 50(5), 1375–1389.
}
\seealso{
dmatrix, smatrix
}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
