% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{landau_index}
\alias{landau_index}
\title{Landau index}
\usage{
landau_index(dmatrix)
}
\arguments{
\item{dmatrix}{dyadic matrix}
}
\value{
h index
}
\description{
Function to obtain the linearity index developed by Landau (1951).
}
\details{
The function landau_index is obtained by the following expression: h = (12/n^3-n)* sum(Va - ((n-1)/2))^2,
where "h" is the linearity index, "n" is the total of animals, "Va" is the total of
times that animal "i" dominated other animals.
}
\references{
Landau, H. G. (1951). On dominance relations and the structure of animal
societies: I. Effect of inherent characteristics. Bulletin of Mathematical
Biophysics, 13, 1-19.
}
\seealso{
dmatrix
}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
