% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociogram.R
\name{reactorSociogram}
\alias{reactorSociogram}
\title{Sociogram plot with reactor information}
\usage{
reactorSociogram(smatrix)
}
\arguments{
\item{smatrix}{sociomatrix}
}
\value{
Circular plot with reactor information
}
\description{
Function to obtain the circle diagram with reactor relationship from an sociomatrix.
}
\details{
The function reactorCircleDiagram is obtained by the sociometric matrix.
Return a circular diagram with reactor relationship between the animals, where
the arrow shows the direction of the relationship and the line thickness
indicates the number of encounters (the thicker the line, the greater
the number of encounters).
}
\examples{
"There is no example"

NULL

}
\author{
Julia P. S. Valente, Matheus Deniz, Karolini T. de Sousa.
}
